Use pest instead of rust-peg
The code is ugly, needs refactoring, but it works! And it's faster!

BUT, while optimizing this code, I realized I was doing something really
slow with find the line and column indices for each atom. That was
probably a big part of the slowness with my rust-peg implementation. So
I'd like to go back to the rust-peg implementation and see if I can make
some optimizations to get the speed closer to my implementation using pest.